1. check GLUT library 2. fix x64 build warnings

+++ src/priv/tool/d3d9Wrapper/myvtxbuf9.cpp
@@ -41,7 +41,7 @@ HRESULT MyVtxBuf9::Lock( UINT OffsetToLock, UINT SizeToLock, VOID ** ppbData, DW
     }
 
     mLockOffset = OffsetToLock;
-    mLockBytes  = ( 0 == SizeToLock && 0 == OffsetToLock ) ? mSysCopy.size() : SizeToLock;
+    mLockBytes  = ( 0 == SizeToLock && 0 == OffsetToLock ) ? (UINT)mSysCopy.size() : SizeToLock;
     mLockFlags  = Flags;
 
     *ppbData = mSysCopy.cptr() + mLockOffset;
